refactor(remediationorchestrator): extract audit_events.go from reconciler.go
Phase 3 file split: moves all 14 emit*Audit methods (DD-AUDIT-003 event
emission for created/workflow-created/EA-created/lifecycle-started/
verifying-started/verification-completed/verification-timed-out/
phase-transition/completion/failure/routing-blocked/approval-requested/
timeout/retention-cleanup) plus resolveWorkflowDisplay into their own file.
Pure structural move: 376 Ginkgo specs still pass.
